# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechXayan Creative Video Editor API
  version: 1.0.0
  description: |
    Production-grade enterprise backend API for TechXayan Creative / my_editor (Windows + Android).
    Provides authentication, project sync with optimistic concurrency control, media asset ingestion
    and intelligence (vector/semantic search), provider-agnostic AI gateway, asynchronous BullMQ job processing,
    speech-to-text with word-level timing, natural-language editor command orchestration,
    realtime collaboration & review timecode commenting, WebSocket/SSE domain event streaming,
    HMAC-verified webhook ingestion with dead-letter queue, and Prometheus observability metrics.
servers:
  - url: http://localhost:4000
    description: Local Development / Test Server
  - url: https://api.editor.techxayan.com
    description: Production Cluster
paths:
...
  /v1/auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Token refreshed successfully
...
  /v1/projects:
    get:
      summary: List user projects with pagination and search
      operationId: listProjects
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: q
          in: query
          description: Search term across title, description, and tags
          schema: { type: string }
        - name: search
          in: query
          description: Alias for search term q
          schema: { type: string }
        - name: status
          in: query
          description: Filter by status
          schema: { type: string, enum: [active, archived, deleted] }
        - name: includeArchived
          in: query
          description: Whether to include archived projects in results
          schema: { type: boolean, default: false }
        - name: sortBy
          in: query
          description: Field to sort by
          schema: { type: string, enum: [updatedAt, createdAt, title], default: updatedAt }
        - name: sortOrder
          in: query
          description: Sort direction
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: Paginated projects list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProjectResponse'
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      totalPages: { type: integer }
    post:
      summary: Create new project
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectPayload'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
...
  /v1/projects/{id}/archive:
    post:
      summary: Archive project (soft retention)
      operationId: archiveProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Project archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /v1/projects/{id}/restore:
    post:
      summary: Restore archived project back to active state
      operationId: restoreProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Project restored to active status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
...
  /v1/projects/{id}/snapshots:
    post:
      summary: Create manual named backup snapshot
      operationId: createProjectSnapshot
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changeSummary: { type: string, example: "Manual pre-render backup" }
      responses:
        '201':
          description: Backup snapshot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBackupSnapshot'
    get:
      summary: List backup snapshots for project
      operationId: listProjectSnapshots
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of backup snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectBackupSnapshot'
...
  /v1/media/{id}:
    get:
      summary: Retrieve media asset by ID
      description: Returns full asset entity including variants (original, thumbnail, waveform, proxy), technical telemetry, and active presigned download URL.
      operationId: getMediaById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Media asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
        '404':
          description: Asset not found or deleted

    delete:
      summary: Delete media asset (Soft or Permanent)
      description: By default soft-deletes asset (marking status DELETED). Pass permanent=true to purge physical files from S3/MinIO.
      operationId: deleteMedia
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: permanent
          in: query
          description: Purge physical storage binaries immediately
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Media asset deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      status: { type: string, enum: [DELETED] }
                      permanent: { type: boolean }
...
  /v1/media/{id}/archive:
    post:
      summary: Archive media asset
      description: Sets media asset status to ARCHIVED, hiding it from default active bin views while preserving storage.
      operationId: archiveMedia
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Media asset archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
    MediaAsset:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        projectId: { type: string, nullable: true }
        folderId: { type: string, nullable: true }
        name: { type: string }
        originalFilename: { type: string }
        category: { type: string, enum: [video, audio, image, font, lut, sticker, template] }
        fileKey: { type: string }
        mimeType: { type: string }
        fileSizeBytes: { type: integer }
        checksumSha256: { type: string }
        durationSeconds: { type: number }
        width: { type: integer }
        height: { type: integer }
        framerate: { type: number }
        codec: { type: string }
        bitrateKbps: { type: integer }
        audioChannels: { type: integer }
        audioSampleRate: { type: integer }
        rotation: { type: integer }
        orientation: { type: string, enum: [landscape, portrait, square] }
        status: { type: string, enum: [UPLOADING, UPLOADED, PROCESSING, READY, FAILED, DELETING, DELETED, ARCHIVED] }
        isFavorite: { type: boolean }
        downloadUrl: { type: string, format: uri }
        thumbnailUrl: { type: string, format: uri }
        variants:
          $ref: '#/components/schemas/MediaVariants'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MediaVariants:
      type: object
      properties:
        original:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            sizeBytes: { type: integer }
            mimeType: { type: string }
        thumbnail:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            width: { type: integer }
            height: { type: integer }
        waveform:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            peaks:
              type: array
              items: { type: number }
            channels: { type: integer }
            samplesPerPixel: { type: integer }
        proxy:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            resolution: { type: string }
            codec: { type: string }
            sizeBytes: { type: integer }
        previewDerivative:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            resolution: { type: string }
            durationSeconds: { type: number }
...
    CreateProjectPayload:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        canvas:
          $ref: '#/components/schemas/Canvas'
        resolution:
          type: object
          description: "Flutter ProjectModel compatible resolution"
          properties:
            width: { type: integer, example: 1920 }
            height: { type: integer, example: 1080 }
        frameRate: { type: number, example: 30.0 }
        durationMs: { type: integer, example: 60000 }
        tracks:
          type: array
          description: "Flutter ProjectModel compatible top-level track list"
          items:
            $ref: '#/components/schemas/TimelineTrack'
...
    ProjectResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        version: { type: integer }
        schemaVersion: { type: integer, example: 1 }
        resolution:
          type: object
          description: "Flutter ProjectModel compatible resolution"
          properties:
            width: { type: integer }
            height: { type: integer }
        frameRate: { type: number, example: 30.0 }
        durationMs: { type: integer, example: 60000 }
        canvas:
          $ref: '#/components/schemas/Canvas'
        timeline: { type: object }
        tracks:
          type: array
          description: "Flutter ProjectModel compatible tracks with millisecond offsets"
          items:
            $ref: '#/components/schemas/TimelineTrack'
        assets:
          type: array
          items:
            $ref: '#/components/schemas/ProjectAsset'
        markers:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              positionMs: { type: integer }
              label: { type: string }
              color: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Canvas:
      type: object
      properties:
        resolutionWidth: { type: integer, default: 1920 }
        resolutionHeight: { type: integer, default: 1080 }
        framerate: { type: number, default: 30.0 }
        aspectRatio: { type: string, default: "16:9" }
        colorSpace: { type: string, default: "rec709" }
        backgroundColor: { type: string, default: "#000000" }
    ProjectAsset:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [video, audio, image, font, lut, other] }
        uri: { type: string }
        duration: { type: number }
        thumbnailUrl: { type: string }
...
    TimelineTrack:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [video, audio, text, effect, image, overlay] }
        name: { type: string }
        isLocked: { type: boolean, default: false }
        isMuted: { type: boolean, default: false }
        clips:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              assetId: { type: string }
              timelineStartMs: { type: integer }
              durationMs: { type: integer }
              sourceInPointMs: { type: integer }
              sourceOutPointMs: { type: integer }
              speed: { type: number, default: 1.0 }
              volume: { type: number, default: 1.0 }
              isLocked: { type: boolean, default: false }
              isMuted: { type: boolean, default: false }
              text:
                type: object
                properties:
                  content: { type: string }
                  fontFamily: { type: string }
                  fontSize: { type: number }
                  color: { type: string }
                  bold: { type: boolean }
                  italic: { type: boolean }
                  alignment: { type: string, enum: [left, center, right] }
                  positionX: { type: number }
                  positionY: { type: number }
              effects:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    type: { type: string }
                    name: { type: string }
                    isEnabled: { type: boolean }
                    parameters: { type: object }
              keyframes:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    property: { type: string }
                    timeMs: { type: integer }
                    value: { type: number }
                    easing: { type: string }
              audio:
                type: object
                properties:
                  volume: { type: number }
                  pan: { type: number }
                  fadeInMs: { type: integer }
                  fadeOutMs: { type: integer }
                  isMuted: { type: boolean }
                  pitch: { type: number }
                  eq: { type: object }
              captions:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    startMs: { type: integer }
                    endMs: { type: integer }
                    text: { type: string }
                    speaker: { type: string }
...
    ProjectBackupSnapshot:
      type: object
      description: "Point-in-time snapshot of complete project state"
      properties:
        id: { type: string }
        projectId: { type: string, format: uuid }
        versionNumber: { type: integer }
        changeSummary: { type: string }
        isAutoSave: { type: boolean }
        deviceName: { type: string }
        createdAt: { type: string, format: date-time }